Move error files.
