+++ nextgen/query/src/lib.rs
@@ -1,10 +1,10 @@
 mod builder;
-mod errors;
 mod parser;
+mod query_error;
 
 pub use builder::*;
-pub use errors::QueryError;
 pub use parser::*;
+pub use query_error::QueryError;
 
 pub trait Queryable {
     fn matches_criteria(&self, criteria: &Criteria) -> Result<bool, QueryError>;
