+++ nextgen/query/src/query_error.rs
